refactor: Moved utils into summariser as not reused elsewhere.
