--- tilde/utils/__init__.py
@@ -1,2 +1,0 @@
-from tilde.utils.utils import rank_sentences_for_relevance, ranked_sentences_to_summary_with_redundancy_detection
-from tilde.utils.components import TextSegmenter, TopicExtractor, RedundancyDetector, C99Segmenter, FastRAKE, SimpleRTE
