--- tilde/utils/components/__init__.py
@@ -1,4 +1,0 @@
-from tilde.utils.components.base_components import TextSegmenter, TopicExtractor, RedundancyDetector
-from tilde.utils.components.C99Segmenter import C99Segmenter
-from tilde.utils.components.FastRAKE import FastRAKE
-from tilde.utils.components.SimpleRTE import SimpleRTE
